# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{userId}/stripe-customer:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string } }
    put:
      operationId: adminSetUserStripeCustomer
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stripeCustomerId: { type: string, nullable: true }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [stripeCustomerId]
                properties:
                  stripeCustomerId: { type: string, nullable: true }
  /admin/company:
    get:
      operationId: adminGetCompany
      tags: [admin]
      responses:
        "200":
          description: Company branding settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanySettingsResponse" }
    put:
      operationId: adminUpdateCompany
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCompanySettingsBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanySettingsResponse" }
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: All users with progress and role (admin/team only)
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUserRow" }
    post:
      operationId: adminCreateUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminCreateUserBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /admin/users/{userId}:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string } }
    get:
      operationId: adminGetUser
      tags: [admin]
      responses:
        "200":
          description: Full user detail with enrollments and activity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserDetail" }
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUpdateUserBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      responses:
        "204": { description: Deleted }
  /admin/users/{userId}/activity:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string } }
      - { name: limit, in: query, schema: { type: integer, default: 50 } }
      - { name: offset, in: query, schema: { type: integer, default: 0 } }
      - { name: action, in: query, schema: { type: string } }
    get:
      operationId: adminGetUserActivity
      tags: [admin]
      responses:
        "200":
          description: Paginated activity log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActivityPage" }
  /admin/users/{userId}/reset-password:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string } }
    post:
      operationId: adminResetPassword
      tags: [admin]
      responses:
        "200":
          description: Reset link generated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResetPasswordResponse" }
  /admin/users/{userId}/access:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string } }
    put:
      operationId: adminSetUserAccess
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetUserAccessBody" }
      responses:
        "200":
          description: Updated access
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CourseAccess" }
  /admin/team:
    get:
      operationId: adminListTeam
      tags: [admin]
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: adminInviteTeamMember
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteTeamMemberBody" }
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
...
components:
  schemas:
...
    Role:
      type: string
      enum: [super_admin, admin, team, student]
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/Role" }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
...
    CompanySettings:
      type: object
      required: [name]
      properties:
        name: { type: string }
        legalName: { type: string, nullable: true }
        supportEmail: { type: string, nullable: true }
        website: { type: string, nullable: true }
        taxId: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        country: { type: string, nullable: true }
        invoiceFooter: { type: string, nullable: true }
    CompanySettingsResponse:
      type: object
      required: [settings, stripeConfigured]
      properties:
        settings: { $ref: "#/components/schemas/CompanySettings" }
        stripeConfigured: { type: boolean }
    UpdateCompanySettingsBody:
      allOf:
        - $ref: "#/components/schemas/CompanySettings"
...
    AdminUserRow:
      type: object
      required: [id, name, email, role, createdAt, completionPct, lessonsCompleted]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        avatarUrl: { type: string, nullable: true }
        role: { $ref: "#/components/schemas/Role" }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        completionPct: { type: number }
        lessonsCompleted: { type: integer }
        coursesEnrolled: { type: integer }
    AdminUserDetail:
      type: object
      required: [user, courseAccess, recentActivity]
      properties:
        user: { $ref: "#/components/schemas/User" }
        courseAccess:
          type: array
          items: { $ref: "#/components/schemas/CourseAccess" }
        recentActivity:
          type: array
          items: { $ref: "#/components/schemas/ActivityEntry" }
        completionPct: { type: number }
        totalLessons: { type: integer }
        completedLessons: { type: integer }
        activitySummary: { $ref: "#/components/schemas/ActivitySummary" }
    CourseAccess:
      type: object
      required: [courseId, courseTitle, hasAccess, completedLessons, totalLessons]
      properties:
        courseId: { type: string }
        courseTitle: { type: string }
        hasAccess: { type: boolean }
        completedLessons: { type: integer }
        totalLessons: { type: integer }
    ActivityEntry:
      type: object
      required: [id, action, target, createdAt]
      properties:
        id: { type: string }
        action: { type: string }
        target: { type: string }
        path: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt: { type: string, format: date-time }
    ActivityPage:
      type: object
      required: [total, items]
      properties:
        total: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/ActivityEntry" }
    ActivitySummary:
      type: object
      required:
        - totalEvents
        - lessonsCompleted
        - totalLessons
        - completionPct
        - downloads
        - videoCompletions
        - refundRequests
        - recentActions
      properties:
        totalEvents: { type: integer }
        lessonsCompleted: { type: integer }
        totalLessons: { type: integer }
        completionPct: { type: integer }
        downloads: { type: integer }
        videoCompletions: { type: integer }
        refundRequests: { type: integer }
        firstSeenAt: { type: string, format: date-time, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        furthestChapter:
          type: object
          nullable: true
          required: [courseTitle, chapterTitle, chapterIndex]
          properties:
            courseTitle: { type: string }
            chapterTitle: { type: string }
            chapterIndex: { type: integer }
        recentActions:
          type: array
          items:
            type: object
            required: [action, target, at]
            properties:
              action: { type: string }
              target: { type: string }
              at: { type: string, format: date-time }
...
    AdminCreateUserBody:
      type: object
      required: [email, name]
      properties:
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    AdminUpdateUserBody:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    SetUserAccessBody:
      type: object
      required: [courseIds]
      properties:
        courseIds:
          type: array
          items: { type: string }
    InviteTeamMemberBody:
      type: object
      required: [email, name, role]
      properties:
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    ResetPasswordResponse:
      type: object
      required: [resetUrl]
      properties:
        resetUrl: { type: string }
        message: { type: string }
...